requirements.txt
streamlit==1.32.0
langchain==0.1.9
langchain-community==0.0.19
transformers==4.37.2
torch==2.1.2
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
scikit-learn==1.3.2
python-dotenv==1.0.0
mysql-connector-python==8.3.0
sqlalchemy==2.0.27
pyyaml==6.0.1
accelerate==0.25.0
bitsandbytes==0.41.3


# cd /mnt/d/DS\ Northeastern/DS\ 5983\ -\ LLMS/data-analysis-llm-agent/data-analysis-llm-agent
# source venv/bin/activate

